// frontend/src/types/auction.d.ts
export interface Auction {
  docid: string;
  boCd: string;
  saNo: string;
  maemulSer: string;
  mokmulSer: string;
  srnSaNo: string;
  jpDeptCd: string;
  jinstatCd: string;
  mulStatcd: string;
  mulJinYn: string;
  maemulUtilCd: string;
  mulBigo: string;
  gamevalAmt: string;
  minmaePrice: string;
  yuchalCnt: string;
  maeAmt: string;
  inqCnt: string;
  gwansMulRegCnt: string;
  remaeordDay: string;
  ipchalGbncd: string;
  maeGiil: string;
  maegyuljGiil: string;
  maeHh1: string;
  maeHh2: string;
  maeHh3: string;
  maeHh4: string;
  notifyMinmaePrice1: string;
  notifyMinmaePrice2: string;
  notifyMinmaePrice3: string;
  notifyMinmaePrice4: string;
  notifyMinmaePriceRate1: string;
  notifyMinmaePriceRate2: string;
  maeGiilCnt: string;
  ipgiganFday: string;
  ipgiganTday: string;
  maePlace: string;
  spJogCd: string;
  mokGbncd: string;
  jongCd: string;
  stopsaGbncd: string;
  daepyoSidoCd: string;
  daepyoSiguCd: string;
  daepyoDongCd: string;
  daepyoRdCd: string;
  hjguSido: string;
  hjguSigu: string;
  hjguDong: string;
  hjguRd: string;
  daepyoLotno: string;
  buldNm: string;
  buldList: string;
  areaList: string;
  jimokList: string;
  lclsUtilCd: string;
  mclsUtilCd: string;
  sclsUtilCd: string;
  jejosaNm: string;
  fuelKindcd: string;
  bsgFormCd: string;
  carNm: string;
  carYrtype: string;
  xCordi: string;
  yCordi: string;
  cordiLvl: string;
  bgPlaceSidoCd: string;
  bgPlaceSiguCd: string;
  bgPlaceDongCd: string;
  bgPlaceRdCd: string;
  bgPlaceLotno: string;
  bgPlaceSido: string;
  bgPlaceSigu: string;
  bgPlaceDong: string;
  bgPlaceRd: string;
  srchHjguBgFlg: string;
  pjbBuldList: string;
  minArea: string;
  maxArea: string;
  groupmaemulser: string;
  bocdsano: string;
  dupSaNo: string;
  byungSaNo: string;
  srchLclsUtilCd: string;
  srchMclsUtilCd: string;
  srchSclsUtilCd: string;
  srchHjguSidoCd: string;
  srchHjguSiguCd: string;
  srchHjguDongCd: string;
  srchHjguRdCd: string;
  srchHjguLotno: string;
  jiwonNm: string;
  jpDeptNm: string;
  tel: string;
  maejibun: string;
  wgs84Xcordi: string;
  wgs84Ycordi: string;
  rd1Cd: string;
  rd2Cd: string;
  rd3Rd4Cd: string;
  rd1Nm: string;
  rd2Nm: string;
  rdEubMyun: string;
  rdNm: string;
  buldNo: string;
  rdAddrSub: string;
  addrGbncd: string;
  bgPlaceRdAllAddr: string;
  bgPlaceAddrGbncd: string;
  srchRd1Cd: string;
  srchRd2Cd: string;
  srchRd3Rd4Cd: string;
  alias: string;
  dummyField: string;
  dspslUsgNm: string;
  convAddr: string;
  printSt: string;
  printCsNo: string;
  colMerge: string;
}